+++ Basis/Packages/dev.hai-vr.basis.comms/Scripts/Systems/Runtime/ScriptableObjects/HVRAddress.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+namespace HVR.Basis.Comms
+{
+    [CreateAssetMenu(fileName = "HVRAddress", menuName = "HVR/Address")]
+    public class HVRAddress : ScriptableObject
+    {
+        public string path;
+    }
+}