+++ Basis/Packages/dev.hai-vr.basis.comms/CHANGELOG.md
@@ -8,13 +8,14 @@ New additions:
 - The toggles are accessible in-game through "Settings > My Avatar" if at least once VixxyMenuItem component is present on the avatar.
 
 Modifications in Basis SDK:
-- The following types are now allowed in avatars: `HVR.Vixxy.HVRVixxyControl`, `HVR.Vixxy.HVRVixxyMenuItem`, and `HVR.Vixxy.HVRVixxyAggregator`.
+- The following types are now allowed in avatars: HVR.Vixxy.HVRVixxyControl, HVR.Vixxy.HVRVixxyMenuItem, and HVR.Vixxy.HVRVixxyAggregator.
 
 Modifications in existing HVR.Basis systems:
 - Acquisition Service now keeps track of values that were sent to it, rather than just emitting events.
 - The SettingsProvider avatar tab override now integrates with Vixxy menu items.
 - HVRAvatarComms now uses the networking carrier at index 1 (this means the second one) for networking Vixxy-specific messages.
 - Remove "Avatar" from the display name of the package, so that is is now "HVR Basis Comms".
+- Renamed HVRAddress to HVRAddressRegistry, so that the name HVRAddress can be used for a new addition.
 
 Editor modifications in existing HVR.Basis systems:
 - Most HVR components now show additional information in Play Mode about their internal state.